feat: show validation errors on each step
