+++ src/components/steps/LoggerStep.tsx
@@ -8,6 +8,8 @@ import { DatePicker } from '../ui/date-picker';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../ui/select';
 import { Textarea } from '../ui/textarea';
 import type { IEATask43Schema, LoggerOEM } from '../../types/schema';
+import { StepValidation } from '../StepValidation';
+import { validateLoggers } from '../../utils/validations';
 import DynamicLoggerOptionalFields from './DynamicLoggerOptionalFields';
 
 export function LoggerStep() {
@@ -17,6 +19,8 @@ export function LoggerStep() {
     locations.reduce((acc, loc) => ({ ...acc, [loc.uuid]: true }), {})
   );
   const [expandedLoggers, setExpandedLoggers] = useState<{ [key: number]: boolean }>({});
+  const formData = watch();
+  const { issues } = validateLoggers(formData);
 
   const addLogger = (locationIndex: number) => {
     const currentLoggers = watch(`measurement_location.${locationIndex}.logger_main_config`) || [];
@@ -57,6 +61,7 @@ export function LoggerStep() {
 
   return (
     <div className="space-y-8">
+      <StepValidation issues={issues} />
       <h2 className="text-2xl font-bold text-primary mb-2">Logger Configuration</h2>
       <div className="text-muted-foreground mb-6">
         <p>A separate logger file is required for each data file which will be uploaded to the system. Data files should ensure consistency in timestamp conventions, averaging periods, etc. for all parameters contained within those files – care should be taken that this is the case when data files contain outputs from multiple sensors.</p>
